+++ R/plot_sobol.R
@@ -17,15 +17,17 @@ theme_AP <- function() {
 ##################################################################################
 #' Visualization of first, total, second and third order Sobol' indices.
 #'
-#' It plots first, total, second and third order Sobol' indices.
+#' It plots first, total, second and third-order Sobol' indices.
 #'
-#' @param data The output of \code{\link{sobol_indices}}.
+#' @param x The output of \code{\link{sobol_indices}}.
 #' @param order If \code{order = "first"}, it plots first and total-order effects.
 #' If \code{order = "second"}, it plots second-order effects. If \code{order = "third"}, it plots
 #' third-order effects. Default is \code{order = "first"}.
 #' @param dummy The output of \code{\link{sobol_dummy}}. Default is NULL.
+#' @param ... Other graphical parameters to plot.
 #'
 #' @return A \code{ggplot} object.
+#' @rdname plot.sensobol
 #' @import ggplot2
 #' @export
 #'
@@ -43,9 +45,94 @@ theme_AP <- function() {
 #' ind <- sobol_indices(Y = Y, N = N, params = params, boot = TRUE, R = R)
 #'
 #' # Plot Sobol' indices
-#' plot_sobol(data = ind)
+#' plot(ind)
 
+plot.sensobol <- function(x, order = "first", dummy = NULL, ...) {
+  sensitivity <- parameters <- original <- low.ci <- high.ci <- NULL
+  data <- x$results
+  colNames <- colnames(data)
+
+  if (order == "first") {
+    dt <- data[sensitivity %in% c("Si", "Ti")]
+    gg <- ggplot2::ggplot(dt, ggplot2::aes(parameters, original, fill = sensitivity)) +
+      ggplot2::geom_bar(stat = "identity",
+                        position = ggplot2::position_dodge(0.6),
+                        color = "black") +
+      ggplot2::scale_y_continuous(breaks = scales::pretty_breaks(n = 3)) +
+      ggplot2::labs(x = "",
+                    y = "Sobol' index") +
+      ggplot2::scale_fill_discrete(name = "Sobol' indices",
+                                   labels = c(expression(S[italic(i)]),
+                                              expression(T[italic(i)]))) +
+      theme_AP()
+
+    if (any(grepl("high.ci", colNames)) == TRUE) {
+      gg <- gg +
+        ggplot2::geom_errorbar(ggplot2::aes(ymin = low.ci,
+                                            ymax = high.ci),
+                               position = ggplot2::position_dodge(0.6))
+    }
+
+    if (is.null(dummy) == FALSE) {
+      col_names <- colnames(dummy)
+
+      if(any(grepl("high.ci", col_names)) == TRUE) {
+        lmt <- dummy$high.ci
+
+      } else {
+        lmt <- dummy$original
+      }
+      gg <- gg +
+        ggplot2::geom_hline(data = dummy,
+                            ggplot2::aes(yintercept = lmt, color = sensitivity),
+                            lty = 2) +
+        ggplot2::guides(linetype = FALSE, color = FALSE)
+    }
+
+  } else if (!order == "first") {
+
+    if (order == "second") {
+      dt <- data[sensitivity %in% "Sij"][low.ci > 0]
+
+    } else if (order == "third") {
+      dt <- data[sensitivity %in% "Sijk"][low.ci > 0]
+
+    } else {
+      stop("Order should be first, second or third")
+    }
+    gg <- ggplot2::ggplot(dt, ggplot2::aes(stats::reorder(parameters, original),
+                                           original)) +
+      ggplot2::geom_point() +
+      ggplot2::geom_errorbar(ggplot2::aes(ymin = low.ci,
+                                          ymax = high.ci)) +
+      ggplot2::labs(x = "",
+                    y = "Sobol' index") +
+      ggplot2::geom_hline(yintercept = 0,
+                          lty = 2,
+                          color = "red") +
+      theme_AP()
+  }
+  return(gg)
+}
+
+
+#' Visualization of first, total, second and third order Sobol' indices (Deprecated,
+#' substituted by plot.sensobol).
+#'
+#' It plots first, total, second and third-order Sobol' indices.
+#'
+#' @param data The output of \code{\link{sobol_indices}}.
+#' @param order If \code{order = "first"}, it plots first and total-order effects.
+#' If \code{order = "second"}, it plots second-order effects. If \code{order = "third"}, it plots
+#' third-order effects. Default is \code{order = "first"}.
+#' @param dummy The output of \code{\link{sobol_dummy}}. Default is NULL.
+#'
+#' @return A \code{ggplot} object.
+#' @import ggplot2
+#' @export
 plot_sobol <- function(data, order = "first", dummy = NULL) {
+  .Deprecated("This function is substituted by the S3 function plot.sensobol and
+              will be removed from future versions")
   sensitivity <- parameters <- original <- low.ci <- high.ci <- NULL
   data <- data$results
   colNames <- colnames(data)
@@ -112,6 +199,7 @@ plot_sobol <- function(data, order = "first", dummy = NULL) {
   }
   return(gg)
 }
+
 
 # PLOT MODEL OUTPUT UNCERTAINTY
 ##################################################################################